requirements.txt
azure-cognitiveservices-speech>=1.35.0
azure-ai-language-conversations>=1.1.0
azure-ai-language-questionanswering>=1.1.0
azure-core>=1.30.0
python-dotenv>=1.0.0
pytest>=8.0
pytest-timeout>=2.2
